fix: Merge the local api with the local v1 api
